+++ test/testSchemaGenerator.js
@@ -620,6 +620,91 @@ describe('generating schema from shorthand', () => {
     assert.doesNotThrow(makeExecutableSchema.bind(null, { typeDefs: short, resolvers: rf, resolverValidationOptions: { requireResolversForNonScalar: false } }), SchemaError); // eslint-disable-line max-len
   });
 
+  it('throws if conflicting validation options are passed', () => {
+    const typeDefs = `
+    type Bird {
+      id: ID
+    }
+    type Query {
+      bird: Bird
+    }
+    schema {
+      query: Query
+    }`;
+    const resolvers = {};
+
+    function assertOptionsError(resolverValidationOptions) {
+      assert.throws(() => makeExecutableSchema({ typeDefs, resolvers, resolverValidationOptions }), TypeError); // eslint-disable-line max-len
+    }
+
+    assertOptionsError({
+      requireResolversForAllFields: true,
+      requireResolversForNonScalar: false,
+      requireResolversForArgs: false,
+    });
+    assertOptionsError({
+      requireResolversForAllFields: true,
+      requireResolversForNonScalar: false,
+    });
+    assertOptionsError({
+      requireResolversForAllFields: true,
+      requireResolversForArgs: false,
+    });
+  });
+
+  it('throws for any missing field if `resolverValidationOptions.requireResolversForAllFields` = true', () => { // eslint-disable-line max-len
+    const typeDefs = `
+    type Bird {
+      id: ID
+    }
+    type Query {
+      bird: Bird
+    }
+    schema {
+      query: Query
+    }`;
+
+    function assertFieldError(errorMatcher, resolvers) {
+      assert.throws(() => makeExecutableSchema({ typeDefs, resolvers, resolverValidationOptions: { requireResolversForAllFields: true } }), SchemaError, errorMatcher); // eslint-disable-line max-len
+    }
+
+    assertFieldError(null, {});
+    assertFieldError('Query.bird', {
+      Bird: {
+        id: bird => bird.id,
+      },
+    });
+    assertFieldError('Bird.id', {
+      Query: {
+        bird: () => ({ id: '123' }),
+      },
+    });
+  });
+
+  it('does not throw if all fields are satisfied when `resolverValidationOptions.requireResolversForAllFields` = true', () => { // eslint-disable-line max-len
+    const typeDefs = `
+    type Bird {
+      id: ID
+    }
+    type Query {
+      bird: Bird
+    }
+    schema {
+      query: Query
+    }`;
+
+    const resolvers = {
+      Bird: {
+        id: bird => bird.id,
+      },
+      Query: {
+        bird: () => ({ id: '123' }),
+      },
+    };
+
+    assert.doesNotThrow(() => makeExecutableSchema({ typeDefs, resolvers, resolverValidationOptions: { requireResolversForAllFields: true } })); // eslint-disable-line max-len
+  });
+
   it('throws an error if a resolve field cannot be used', (done) => {
     const shorthand = `
       type BirdSpecies {
@@ -743,6 +828,38 @@ describe('providing useful errors from resolve functions', () => {
     graphql(jsSchema, testQuery).then((res) => {
       assert.equal(logger.errors.length, 1);
       assert.match(logger.errors[0].message, expectedErr);
+      assert.deepEqual(res.data, expectedResData);
+      done();
+    });
+  });
+
+  it('will not throw errors on undefined by default', (done) => {
+    const shorthand = `
+      type RootQuery {
+        species(name: String): String
+        stuff: String
+      }
+      schema {
+        query: RootQuery
+      }
+    `;
+    const resolve = {
+      RootQuery: {
+        species: () => undefined,
+        stuff: () => 'stuff',
+      },
+    };
+
+    const logger = new Logger();
+    const jsSchema = makeExecutableSchema({
+      typeDefs: shorthand,
+      resolvers: resolve,
+      logger,
+    });
+    const testQuery = '{ species, stuff }';
+    const expectedResData = { species: null, stuff: 'stuff' };
+    graphql(jsSchema, testQuery).then((res) => {
+      assert.equal(logger.errors.length, 0);
       assert.deepEqual(res.data, expectedResData);
       done();
     });